Add more test for target indirection in UPDATE
Queries of the form:

```
UPDATE ... SET (a, b) = (SELECT '1', '2')
UPDATE ... SET (b, a) = (SELECT '2', '1')
```

Should do the same thing, but currently the order of the attributes in
`SET (...)` as rewriten for pushdown is only based on the physical ordering
of the attributes in the relation. This leads to several subtle problems
including situation where a DROPped then reADDed attributes will change its
placement in the attribute list.

There are maybe more tests to add in other situation where a SET
(MULTIEXPR) is possible, though UPDATE form is pretty unique as
alternatives are not supported by citus: `(INSERT .. ON CONFLICT SET (a,b).....`